kernel: flycheck parameters
Add include paths, definitions, and includes used by the kernel
build for flycheck to work correctly on kernel code. This
requires setting the kernel-arch variable as well as pointing the
minor mode at a built kernel tree so it can find generated
headers.

It's still missing a few options and needs some additional
customization, but should work for the majority of kernel code.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>